requirements.txt
Flask==1.0.2

# Application server for both development and production.
gunicorn==19.9.0

# Testing and static analysis.
pytest==4.3.0
pytest-cov==2.6.1
mock==2.0.0
flake8==3.7.6

# CLI.
Click==7.0

# Data and workers.
psycopg2==2.7.7
psycopg2-binary==2.8.2
Flask-SQLAlchemy==2.3.2
redis==3.2.0
celery-once==2.1.0
celery-singleton==0.1.3
# celery==4.1.1
rabbitmq==0.2.0

# Forms.
Flask-WTF==0.9.5
WTForms-Components==0.10.3
wtforms-alchemy==0.16.8
flask-paginate==0.5.2

# Payments.
stripe==2.21.0

# Utils.
fake-factory==9999.9.9
tzlocal==1.5.1
beautifulsoup4==4.7.1
mailgun==0.1.1
flanker==0.9.0
#dateparser==0.7.0
easygui==0.98.1
flask-paginate==0.5.2
# numpy==1.8.0
gdata==2.0.18
amqp==2.4.1
celery==4.2.1
mail-parser==3.9.2
html2text==2018.1.9
emaildata==0.3.4
flask-compress==1.4.0
mailchimp3==3.0.6
cryptography==2.5
simple-crypt==4.1.7
pydrive==1.3.1
airtable==0.3.1
airtable-python-wrapper==0.11.3.post1
pur==5.2.1
Werkzeug==0.14.1
celery-redbeat==0.13.0
repos==0.0.1
pprint==0.1

# APIs

# Extensions.
flask-debugtoolbar==0.10.1
Flask-Mail==0.9.1
Flask-Login==0.3.2
Flask-Cache==0.13.1
Flask-Caching==1.4.0

# Threading.
# asyncio==3.4.3
# aiohttp
# asyncio_stripe
# attrs
# aiopg
# websocket-client==0.53.0

# Date
ago==0.0.93

kombu==4.4.0
